feat: Make GenerateSourcePluginDocs struct method
Also, generate README.md with all tables
in addition to file per table.
