+++ plugins/.snapshots/TestGenerateSourcePluginDocs-README.md
@@ -1,0 +1,7 @@
+
+# Source Plugin: test
+## Tables
+| Name          | Description   |
+| ------------- | ------------- |
+|test_table|Description for test table|
+
